# Conflicts:
#	Jenkinsfile
#	backend/de.metas.adempiere.adempiere/base/src/main/java-legacy/org/compiere/swing/ComboBoxAutoCompletion.java
#	backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/dao/impl/InArrayQueryFilter.java
#	backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/sql/impl/AbstractCStatementProxy.java
#	backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/sql/impl/CPreparedStatementProxy.java
#	backend/de.metas.adempiere.adempiere/client/src/main/java-legacy/org/compiere/apps/AMenuStartItem.java
#	backend/de.metas.adempiere.adempiere/client/src/main/java-legacy/org/compiere/grid/VCreateFromFactory.java
#	backend/de.metas.adempiere.adempiere/client/src/main/java/org/compiere/apps/search/InfoSimple.java
#	backend/de.metas.business/src/main/java/de/metas/quantity/Quantitys.java
#	backend/de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/api/impl/InvoiceCandBL.java
#	backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/view/descriptor/SqlViewRowIdsConverters.java
#	backend/de.metas.vertical.pharma/src/main/java/de/metas/impexp/bpartner/IFABPartnerLocationImportHelper.java
#	backend/pom.xml
#	backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_440_request/pom.xml
#	backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_440_response/pom.xml
#	backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_450_request/pom.xml
#	backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_xversion/src/main/java/de/metas/vertical/healthcare_ch/forum_datenaustausch_ch/invoice_xversion/JaxbUtil.java
#	misc/parent-pom/pom.xml
#	misc/services/camel/de-metas-camel-edi/pom.xml

+++ backend/de.metas.shipper.gateway.derkurier/src/main/java/de/metas/shipper/gateway/derkurier/DerKurierDeliveryOrderRepository.java
@@ -250,7 +250,7 @@ public class DerKurierDeliveryOrderRepository implements DeliveryOrderRepository
 					.build();
 			deliveryPositionBuilder.packageDimensions(packageDimensions);
 		}
-		deliveryPositionBuilder.grossWeightKg(lineRecord.getDK_ParcelWeight().intValue());
+		deliveryPositionBuilder.grossWeightKg(lineRecord.getDK_ParcelWeight());
 
 		final DerKurierDeliveryData derKurierDeliveryData = DerKurierDeliveryData.builder()
 				.station(assertSameAsPreviousValue(COLUMNNAME_DK_Consignee_DesiredStation, lineRecord, previousLineRecord))
@@ -411,7 +411,7 @@ public class DerKurierDeliveryOrderRepository implements DeliveryOrderRepository
 				lineRecord.setDK_ParcelLength(BigDecimal.valueOf(packageDimensions.getLengthInCM()));
 				lineRecord.setDK_ParcelWidth(BigDecimal.valueOf(packageDimensions.getWidthInCM()));
 			}
-			lineRecord.setDK_ParcelWeight(BigDecimal.valueOf(deliveryPosition.getGrossWeightKg()));
+			lineRecord.setDK_ParcelWeight(deliveryPosition.getGrossWeightKg());
 
 			lineRecord.setDK_ParcelNumber(derKurierDeliveryData.getParcelNumber());
